requirements.txt
pytorch==1.9.0
torchvision=0.10.0
tensorboard
scikit-learn